requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.5.0.5
celery==4.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==1.8.2
django-cors-middleware==1.3.1
django-haystack==2.8.1
django-redis==4.4.0
django-tinymce==2.6.0
dnspython==1.16.0
eventlet==0.28.0
fdfs-client-py==1.2.6
greenlet==0.4.17
idna==2.10
importlib-metadata==2.0.0
itsdangerous==0.24
jieba==0.42.1
kombu==4.2.0
monotonic==1.5
mutagen==1.40.0
Pillow==7.2.0
prompt-toolkit==3.0.7
PyMySQL==0.10.1
pytz==2020.1
redis==2.10.6
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
Whoosh==2.7.4
zipp==3.3.0
